Move new SinkWrapper related classes to another package
Extend Javadoc of PipelineSink and RandomAccessSink to reference the
more effective SinkWrapper approach
