+++ doxia-core/src/main/java/org/apache/maven/doxia/sink/impl/PipelineSink.java
@@ -28,8 +28,10 @@ import org.apache.maven.doxia.sink.Sink;
 
 /**
  * May be used to invoke the same method on a List of Sinks.
+ * Consider using the more flexible {@link SinkWrapper} approach which doesn't require the use of dynamic proxies.
  *
  * @author <a href="mailto:trygvis@inamo.no">Trygve Laugst&oslash;l</a>
+ * @see SinkWrapper
  */
 public class PipelineSink implements InvocationHandler {
     private List<Sink> pipeline;
